// backend/src/model/challenge.ts
/**
 * apispec
 * EU IV Madness API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Country } from './country';
import { IdeaGroup } from './ideaGroup';
import { Achievement } from './achievement';


/**
 * A challenge
 */
export interface Challenge { 
    achievements: Array<Achievement>;
    ideaGroups: Array<IdeaGroup>;
    country: Country;
}