// backend/src/model/challengeBody.ts
/**
 * apispec
 * EU IV Madness API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChallengeBody { 
    continents: Array<number>;
    ideaGroupCount: number;
    achievementCount: number;
}