// backend/src/model/country.ts
/**
 * apispec
 * EU IV Madness API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * A Country
 */
export interface Country { 
    id: number;
    name: string;
    continent: number;
}